+++ openeo/extra/job_management/__init__.py
@@ -1,11 +1,9 @@
-import abc
 import collections
 import contextlib
 import dataclasses
 import datetime
 import json
 import logging
-import re
 import time
 import warnings
 from pathlib import Path
@@ -14,7 +12,6 @@ from typing import (
     Any,
     Callable,
     Dict,
-    Iterable,
     List,
     Mapping,
     NamedTuple,
@@ -23,12 +20,8 @@ from typing import (
     Union,
 )
 
-import numpy
 import pandas as pd
 import requests
-import shapely.errors
-import shapely.geometry.base
-import shapely.wkt
 from requests.adapters import HTTPAdapter
 from urllib3.util import Retry
 
@@ -37,17 +30,21 @@ from openeo.extra.job_management._thread_worker import (
     _JobManagerWorkerThreadPool,
     _JobStartTask,
 )
-from openeo.internal.processes.parse import (
-    Parameter,
-    Process,
-    parse_remote_process_definition,
-)
+from openeo.extra.job_management.process_based_job_creator import ProcessBasedJobCreator
+from openeo.extra.job_management.dataframe_job_db import JobDatabaseInterface, FullDataFrameJobDatabase, ParquetJobDatabase, CsvJobDatabase
+
 from openeo.rest import OpenEoApiError
 from openeo.rest.auth.auth import BearerAuth
-from openeo.util import LazyLoadCache, deep_get, repr_truncate, rfc3339
+from openeo.util import  deep_get, rfc3339
 
 _log = logging.getLogger(__name__)
 
+__all__ = [
+    "JobDatabaseInterface",
+    "FullDataFrameJobDatabase",
+    "ParquetJobDatabase",
+    "CsvJobDatabase",
+]
 
 class _Backend(NamedTuple):
     """Container for backend info/settings"""
@@ -62,67 +59,6 @@ MAX_RETRIES = 50
 
 # Sentinel value to indicate that a parameter was not set
 _UNSET = object()
-
-
-class JobDatabaseInterface(metaclass=abc.ABCMeta):
-    """
-    Interface for a database of job metadata to use with the :py:class:`MultiBackendJobManager`,
-    allowing to regularly persist the job metadata while polling the job statuses
-    and resume/restart the job tracking after it was interrupted.
-
-    .. versionadded:: 0.31.0
-    """
-
-    @abc.abstractmethod
-    def exists(self) -> bool:
-        """Does the job database already exist, to read job data from?"""
-        ...
-
-    @abc.abstractmethod
-    def persist(self, df: pd.DataFrame):
-        """
-        Store (now or updated) job data to the database.
-
-        The provided dataframe may only cover a subset of all the jobs ("rows") of the whole database,
-        so it should be merged with the existing data (if any) instead of overwriting it completely.
-
-        :param df: job data to store.
-        """
-        ...
-
-    @abc.abstractmethod
-    def count_by_status(self, statuses: Iterable[str] = ()) -> dict:
-        """
-        Retrieve the number of jobs per status.
-
-        :param statuses: List/set of statuses to include. If empty, all statuses are included.
-
-        :return: dictionary with status as key and the count as value.
-        """
-        ...
-
-    @abc.abstractmethod
-    def get_by_status(self, statuses: List[str], max=None) -> pd.DataFrame:
-        """
-        Returns a dataframe with jobs, filtered by status.
-
-        :param statuses: List of statuses to include.
-        :param max: Maximum number of jobs to return.
-
-        :return: DataFrame with jobs filtered by status.
-        """
-        ...
-
-    @abc.abstractmethod
-    def get_by_indices(self, indices: Iterable[Union[int, str]]) -> pd.DataFrame:
-        """
-        Returns a dataframe with jobs based on their (dataframe) index
-
-        :param indices: List of indices to include.
-
-        :return: DataFrame with jobs filtered by indices.
-        """
-        ...
 
 
 def _start_job_default(row: pd.Series, connection: Connection, *args, **kwargs):
@@ -907,197 +843,6 @@ def ignore_connection_errors(context: Optional[str] = None, sleep: int = 5):
         time.sleep(sleep)
 
 
-class FullDataFrameJobDatabase(JobDatabaseInterface):
-    def __init__(self):
-        super().__init__()
-        self._df = None
-
-    def initialize_from_df(self, df: pd.DataFrame, *, on_exists: str = "error"):
-        """
-        Initialize the job database from a given dataframe,
-        which will be first normalized to be compatible
-        with :py:class:`MultiBackendJobManager` usage.
-
-        :param df: dataframe with some columns your ``start_job`` callable expects
-        :param on_exists: what to do when the job database already exists (persisted on disk):
-            - "error": (default) raise an exception
-            - "skip": work with existing database, ignore given dataframe and skip any initialization
-
-        :return: initialized job database.
-
-        .. versionadded:: 0.33.0
-        """
-        # TODO: option to provide custom MultiBackendJobManager subclass with custom normalize?
-        if self.exists():
-            if on_exists == "skip":
-                return self
-            elif on_exists == "error":
-                raise FileExistsError(f"Job database {self!r} already exists.")
-            else:
-                # TODO handle other on_exists modes: e.g. overwrite, merge, ...
-                raise ValueError(f"Invalid on_exists={on_exists!r}")
-        df = MultiBackendJobManager._normalize_df(df)
-        self.persist(df)
-        # Return self to allow chaining with constructor.
-        return self
-
-    @abc.abstractmethod
-    def read(self) -> pd.DataFrame:
-        """
-        Read job data from the database as pandas DataFrame.
-
-        :return: loaded job data.
-        """
-        ...
-
-    @property
-    def df(self) -> pd.DataFrame:
-        if self._df is None:
-            self._df = self.read()
-        return self._df
-
-    def count_by_status(self, statuses: Iterable[str] = ()) -> dict:
-        status_histogram = self.df.groupby("status").size().to_dict()
-        statuses = set(statuses)
-        if statuses:
-            status_histogram = {k: v for k, v in status_histogram.items() if k in statuses}
-        return status_histogram
-
-    def get_by_status(self, statuses, max=None) -> pd.DataFrame:
-        """
-        Returns a dataframe with jobs, filtered by status.
-
-        :param statuses: List of statuses to include.
-        :param max: Maximum number of jobs to return.
-
-        :return: DataFrame with jobs filtered by status.
-        """
-        df = self.df
-        filtered = df[df.status.isin(statuses)]
-        return filtered.head(max) if max is not None else filtered
-
-    def _merge_into_df(self, df: pd.DataFrame):
-        if self._df is not None:
-            unknown_indices = set(df.index).difference(df.index)
-            if unknown_indices:
-                _log.warning(f"Merging DataFrame with {unknown_indices=} which will be lost.")
-            self._df.update(df, overwrite=True)
-        else:
-            self._df = df
-
-    def get_by_indices(self, indices: Iterable[Union[int, str]]) -> pd.DataFrame:
-        indices = set(indices)
-        known = indices.intersection(self.df.index)
-        unknown = indices.difference(self.df.index)
-        if unknown:
-            _log.warning(f"Ignoring unknown DataFrame indices {unknown}")
-        return self._df.loc[list(known)]
-
-
-class CsvJobDatabase(FullDataFrameJobDatabase):
-    """
-    Persist/load job metadata with a CSV file.
-
-    :implements: :py:class:`JobDatabaseInterface`
-    :param path: Path to local CSV file.
-
-    .. note::
-        Support for GeoPandas dataframes depends on the ``geopandas`` package
-        as :ref:`optional dependency <installation-optional-dependencies>`.
-
-    .. versionadded:: 0.31.0
-    """
-
-    def __init__(self, path: Union[str, Path]):
-        super().__init__()
-        self.path = Path(path)
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}({str(self.path)!r})"
-
-    def exists(self) -> bool:
-        return self.path.exists()
-
-    def _is_valid_wkt(self, wkt: str) -> bool:
-        try:
-            shapely.wkt.loads(wkt)
-            return True
-        except shapely.errors.WKTReadingError:
-            return False
-
-    def read(self) -> pd.DataFrame:
-        df = pd.read_csv(
-            self.path,
-            # TODO: possible to avoid hidden coupling with MultiBackendJobManager here?
-            dtype={c: r.dtype for (c, r) in MultiBackendJobManager._COLUMN_REQUIREMENTS.items()},
-        )
-        if (
-            "geometry" in df.columns
-            and df["geometry"].dtype.name != "geometry"
-            and self._is_valid_wkt(df["geometry"].iloc[0])
-        ):
-            import geopandas
-
-            # `df.to_csv()` in `persist()` has encoded geometries as WKT, so we decode that here.
-            df.geometry = geopandas.GeoSeries.from_wkt(df["geometry"])
-            df = geopandas.GeoDataFrame(df)
-        return df
-
-    def persist(self, df: pd.DataFrame):
-        self._merge_into_df(df)
-        self.path.parent.mkdir(parents=True, exist_ok=True)
-        self.df.to_csv(self.path, index=False)
-
-
-class ParquetJobDatabase(FullDataFrameJobDatabase):
-    """
-    Persist/load job metadata with a Parquet file.
-
-    :implements: :py:class:`JobDatabaseInterface`
-    :param path: Path to the Parquet file.
-
-    .. note::
-        Support for Parquet files depends on the ``pyarrow`` package
-        as :ref:`optional dependency <installation-optional-dependencies>`.
-
-        Support for GeoPandas dataframes depends on the ``geopandas`` package
-        as :ref:`optional dependency <installation-optional-dependencies>`.
-
-    .. versionadded:: 0.31.0
-    """
-
-    def __init__(self, path: Union[str, Path]):
-        super().__init__()
-        self.path = Path(path)
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}({str(self.path)!r})"
-
-    def exists(self) -> bool:
-        return self.path.exists()
-
-    def read(self) -> pd.DataFrame:
-        # Unfortunately, a naive `pandas.read_parquet()` does not easily allow
-        # reconstructing geometries from a GeoPandas Parquet file.
-        # And vice-versa, `geopandas.read_parquet()` does not support reading
-        # Parquet file without geometries.
-        # So we have to guess which case we have.
-        # TODO is there a cleaner way to do this?
-        import pyarrow.parquet
-
-        metadata = pyarrow.parquet.read_metadata(self.path)
-        if b"geo" in metadata.metadata:
-            import geopandas
-
-            return geopandas.read_parquet(self.path)
-        else:
-            return pd.read_parquet(self.path)
-
-    def persist(self, df: pd.DataFrame):
-        self._merge_into_df(df)
-        self.path.parent.mkdir(parents=True, exist_ok=True)
-        self.df.to_parquet(self.path, index=False)
-
 
 def get_job_db(path: Union[str, Path]) -> JobDatabaseInterface:
     """
@@ -1140,178 +885,3 @@ def create_job_db(path: Union[str, Path], df: pd.DataFrame, *, on_exists: str = 
     return job_db
 
 
-class ProcessBasedJobCreator:
-    """
-    Batch job creator
-    (to be used together with :py:class:`MultiBackendJobManager`)
-    that takes a parameterized openEO process definition
-    (e.g a user-defined process (UDP) or a remote openEO process definition),
-    and creates a batch job
-    for each row of the dataframe managed by the :py:class:`MultiBackendJobManager`
-    by filling in the process parameters with corresponding row values.
-
-    .. seealso::
-        See :ref:`job-management-with-process-based-job-creator`
-        for more information and examples.
-
-    Process parameters are linked to dataframe columns by name.
-    While this intuitive name-based matching should cover most use cases,
-    there are additional options for overrides or fallbacks:
-
-    -   When provided, ``parameter_column_map`` will be consulted
-        for resolving a process parameter name (key in the dictionary)
-        to a desired dataframe column name (corresponding value).
-    -   One common case is handled automatically as convenience functionality.
-
-        When:
-
-        - ``parameter_column_map`` is not provided (or set to ``None``),
-        - and there is a *single parameter* that accepts inline GeoJSON geometries,
-        - and the dataframe is a GeoPandas dataframe with a *single geometry* column,
-
-        then this parameter and this geometries column will be linked automatically.
-
-    -   If a parameter can not be matched with a column by name as described above,
-        a default value will be picked,
-        first by looking in ``parameter_defaults`` (if provided),
-        and then by looking up the default value from the parameter schema in the process definition.
-    -   Finally if no (default) value can be determined and the parameter
-        is not flagged as optional, an error will be raised.
-
-
-    :param process_id: (optional) openEO process identifier.
-        Can be omitted when working with a remote process definition
-        that is fully defined with a URL in the ``namespace`` parameter.
-    :param namespace: (optional) openEO process namespace.
-        Typically used to provide a URL to a remote process definition.
-    :param parameter_defaults: (optional) default values for process parameters,
-        to be used when not available in the dataframe managed by
-        :py:class:`MultiBackendJobManager`.
-    :param parameter_column_map: Optional overrides
-        for linking process parameters to dataframe columns:
-        mapping of process parameter names as key
-        to dataframe column names as value.
-
-    .. versionadded:: 0.33.0
-
-    .. warning::
-        This is an experimental API subject to change,
-        and we greatly welcome
-        `feedback and suggestions for improvement <https://github.com/Open-EO/openeo-python-client/issues>`_.
-
-    """
-
-    def __init__(
-        self,
-        *,
-        process_id: Optional[str] = None,
-        namespace: Union[str, None] = None,
-        parameter_defaults: Optional[dict] = None,
-        parameter_column_map: Optional[dict] = None,
-    ):
-        if process_id is None and namespace is None:
-            raise ValueError("At least one of `process_id` and `namespace` should be provided.")
-        self._process_id = process_id
-        self._namespace = namespace
-        self._parameter_defaults = parameter_defaults or {}
-        self._parameter_column_map = parameter_column_map
-        self._cache = LazyLoadCache()
-
-    def _get_process_definition(self, connection: Connection) -> Process:
-        if isinstance(self._namespace, str) and re.match("https?://", self._namespace):
-            # Remote process definition handling
-            return self._cache.get(
-                key=("remote_process_definition", self._namespace, self._process_id),
-                load=lambda: parse_remote_process_definition(namespace=self._namespace, process_id=self._process_id),
-            )
-        elif self._namespace is None:
-            # Handling of a user-specific UDP
-            udp_raw = connection.user_defined_process(self._process_id).describe()
-            return Process.from_dict(udp_raw)
-        else:
-            raise NotImplementedError(
-                f"Unsupported process definition source udp_id={self._process_id!r} namespace={self._namespace!r}"
-            )
-
-    def start_job(self, row: pd.Series, connection: Connection, **_) -> BatchJob:
-        """
-        Implementation of the ``start_job`` callable interface
-        of :py:meth:`MultiBackendJobManager.run_jobs`
-        to create a job based on given dataframe row
-
-        :param row: The row in the pandas dataframe that stores the jobs state and other tracked data.
-        :param connection: The connection to the backend.
-        """
-        # TODO: refactor out some methods, for better reuse and decoupling:
-        #       `get_arguments()` (to build the arguments dictionary), `get_cube()` (to create the cube),
-
-        process_definition = self._get_process_definition(connection=connection)
-        process_id = process_definition.id
-        parameters = process_definition.parameters or []
-
-        if self._parameter_column_map is None:
-            self._parameter_column_map = self._guess_parameter_column_map(parameters=parameters, row=row)
-
-        arguments = {}
-        for parameter in parameters:
-            param_name = parameter.name
-            column_name = self._parameter_column_map.get(param_name, param_name)
-            if column_name in row.index:
-                # Get value from dataframe row
-                value = row.loc[column_name]
-            elif param_name in self._parameter_defaults:
-                # Fallback on default values from constructor
-                value = self._parameter_defaults[param_name]
-            elif parameter.has_default():
-                # Explicitly use default value from parameter schema
-                value = parameter.default
-            elif parameter.optional:
-                # Skip optional parameters without any fallback default value
-                continue
-            else:
-                raise ValueError(f"Missing required parameter {param_name!r} for process {process_id!r}")
-
-            # Prepare some values/dtypes for JSON encoding
-            if isinstance(value, numpy.integer):
-                value = int(value)
-            elif isinstance(value, numpy.number):
-                value = float(value)
-            elif isinstance(value, shapely.geometry.base.BaseGeometry):
-                value = shapely.geometry.mapping(value)
-
-            arguments[param_name] = value
-
-        cube = connection.datacube_from_process(process_id=process_id, namespace=self._namespace, **arguments)
-
-        title = row.get("title", f"Process {process_id!r} with {repr_truncate(arguments)}")
-        description = row.get("description", f"Process {process_id!r} (namespace {self._namespace}) with {arguments}")
-        job = connection.create_job(cube, title=title, description=description)
-
-        return job
-
-    def __call__(self, *arg, **kwargs) -> BatchJob:
-        """Syntactic sugar for calling :py:meth:`start_job`."""
-        return self.start_job(*arg, **kwargs)
-
-    @staticmethod
-    def _guess_parameter_column_map(parameters: List[Parameter], row: pd.Series) -> dict:
-        """
-        Guess parameter-column mapping from given parameter list and dataframe row
-        """
-        parameter_column_map = {}
-        # Geometry based mapping: try to automatically map geometry columns to geojson parameters
-        geojson_parameters = [p.name for p in parameters if p.schema.accepts_geojson()]
-        geometry_columns = [i for (i, v) in row.items() if isinstance(v, shapely.geometry.base.BaseGeometry)]
-        if geojson_parameters and geometry_columns:
-            if len(geojson_parameters) == 1 and len(geometry_columns) == 1:
-                # Most common case: one geometry parameter and one geometry column: can be mapped naively
-                parameter_column_map[geojson_parameters[0]] = geometry_columns[0]
-            elif all(p in geometry_columns for p in geojson_parameters):
-                # Each geometry param has geometry column with same name: easy to map
-                parameter_column_map.update((p, p) for p in geojson_parameters)
-            else:
-                raise RuntimeError(
-                    f"Problem with mapping geometry columns ({geometry_columns}) to process parameters ({geojson_parameters})"
-                )
-        _log.debug(f"Guessed parameter-column map: {parameter_column_map}")
-        return parameter_column_map
